frontend

+++ src/main/java/com/example/proj2/models/Gestordeprojeto.java
@@ -12,7 +12,7 @@ import java.math.BigDecimal;
 public class Gestordeprojeto {
     @Id
     @Column(name = "idgestor", nullable = false, precision = 10)
-    private BigDecimal id;
+    private Long id;
 
     @Column(name = "nome", length = 50)
     private String nome;
@@ -26,11 +26,11 @@ public class Gestordeprojeto {
     @Column(name = "password", length = 20)
     private String password;
 
-    public BigDecimal getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigDecimal id) {
+    public void setId(Long id) {
         this.id = id;
     }
